Fix Safe Args type error: use integer instead of long for nav args
- nav_graph.xml: change mementoId and categoryId argType from long to
  integer (Safe Args cannot parse defaultValue="0" as long literal)
- CategoriesFragment: pass category.id.toInt() for categoryId arg
- CategoryDetailFragment: pass args.categoryId.toLong() to ViewModel

+++ app/src/main/java/com/pawel/memento/ui/categories/CategoriesFragment.kt
@@ -30,7 +30,7 @@ class CategoriesFragment : Fragment() {
         adapter = CategoryAdapter(onClick = { category ->
                 findNavController().navigate(
                     CategoriesFragmentDirections.actionCategoriesFragmentToCategoryDetailFragment(
-                        categoryId = category.id,
+                        categoryId = category.id.toInt(),
                         categoryName = category.name
                     )
                 )
